APP-13932: improve tooltips
